Party Afmaken + Verfijnen van code fridge en freezer
het afmaken van de splitbill functie binnen de party mode voor het splitsen van de kostprijs van de drank indien mensen dit wensen te gebruiken ( wel nog de data van card hierinsteken zodat dt optelt)

Hierna het verfijnen van code van partymode fridge en freezer hierbij heb ik gezet dat er bij fridge en freezer een datum picker is en een dropdown menu dit is voor de houdbaarheidsdatum van het product en de categorie ( vlees vis drank enzo) , bij partymode is ook een categorie toegevoeg dus bier zwaar bier sterke drank etc

+++ app/src/main/java/com/example/fridgeapp/Party/CardParty.java
@@ -6,14 +6,14 @@ public class CardParty {
 
     private String title;
     private int quantity;
-    private int price;
+    private double price;
     private String type;
     @Keep
     public CardParty()
     {
         //empty constructor needed
     }
-    public CardParty (String title, int quantity  , String type , int price)
+    public CardParty (String title, int quantity  , String type , double price)
     {
         this.title = title;
         this.quantity = quantity;
@@ -29,7 +29,7 @@ public class CardParty {
         return quantity;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
